asr/requirements.txt
fastapi
requests
uvicorn
transformers
torch
python-multipart
soundfile
scipy
pandas
python-dotenv